Move PluginMessageConnection to PluginMessageHandling
So that other modules (e.g. swift-plugin-server can use it)
Renamed to 'StandardIOMessageConnection'

(cherry picked from commit 3a5ff0e40eb5a99f8d9e5dc2ad99e1414d49cfa5)
