+++ Sources/SwiftCompilerPlugin/CMakeLists.txt
@@ -14,5 +14,4 @@ add_swift_syntax_library(SwiftCompilerPlugin
 target_link_swift_syntax_libraries(SwiftCompilerPlugin PUBLIC
   SwiftSyntaxMacros
   SwiftCompilerPluginMessageHandling
-  _SwiftSyntaxCShims
 )
